+++ src/main/cpp/it/d4np/util/util.hpp
@@ -9,6 +9,7 @@
 
 #include <it/d4np/util/circular_buffer.hpp>
 #include <it/d4np/util/cli_parser.hpp>
+#include <it/d4np/util/file_stream.hpp>
 #include <it/d4np/util/flat_map.hpp>
 #include <it/d4np/util/flat_set.hpp>
 #include <it/d4np/util/hash.hpp>